+++ Matrices.h
@@ -5,17 +5,16 @@
 #include <iostream>
 using namespace std;
 
-void cargarMatriz(vector<vector<int>>& matriz, int filas, int columnas);
-void mostrarMatriz(const vector<vector<int>>& matriz, int filas, int columnas);
-void intercambiarFilas(vector<vector<int>>& matriz, int fila1, int fila2, int columnas);
-void multiplicarFilaPorEscalar(vector<vector<int>>& matriz, int fila, int escalar, int columnas);
+void cargarMatriz(vector<vector<double>>& matriz, double filas, double columnas);
+void mostrarMatriz(const vector<vector<double>>& matriz, double filas, double columnas);
+void intercambiarFilas(vector<vector<double>>& matriz, double fila1, double fila2, double columnas);
+void multiplicarFilaPorEscalar(vector<vector<double>>& matriz, double fila, double escalar, double columnas);
 
-void restarFilas(std::vector<std::vector<int>>& mat, int filaPivote, int filaObjetivo, int factor, int columnas);
+void restarFilas(std::vector<std::vector<double>>& mat, double filaPivote, double filaObjetivo, double factor, double columnas);
 
-int DeterminanteSarrus(const vector<vector<int>> &matriz);
-int DeterminanteLaplace(const vector<vector<int>> &matriz); // recursivo
-vector<vector<int>> menorMatriz(const vector<vector<int>> &matriz, int fila, int col);
-
+double DeterminanteSarrus(const vector<vector<double>> &matriz);
+double DeterminanteLaplace(const vector<vector<double>> &matriz); // recursivo
+vector<vector<double>> menorMatriz(const vector<vector<double>> &matriz, double fila, double col);
