+++ ResDeSist.h
@@ -3,4 +3,4 @@
 #include <vector>
 using namespace std;
 
-vector<int> resolverPorGaussJordan(vector<vector<int>> A, vector<int> b);
+vector<double> resolverPorGaussJordan(vector<vector<double>> A, vector<double> b);
